+++ format/s3m/channel/memory.go
@@ -9,21 +9,6 @@ import (
 	formatutil "github.com/gotracker/playback/util"
 )
 
-type SharedMemory struct {
-	VolSlideEveryFrame  bool
-	LowPassFilterEnable bool
-	// ResetMemoryAtStartOfOrder0 if true will reset the memory registers when the first tick of the first row of the first order pattern plays
-	ResetMemoryAtStartOfOrder0 bool
-	// ST2/Amiga quirks
-	ST2Vibrato          bool
-	ST2Tempo            bool
-	AmigaSlides         bool
-	ZeroVolOptimization bool
-	AmigaLimits         bool
-	// Mod quirks mode
-	ModCompatibility bool
-}
-
 // Memory is the storage object for custom effect/command values
 type Memory struct {
 	portaToNote   memory.Value[DataEffect]
